+++ packages/ai/src/agent/tool-loop-agent.test-d.ts
@@ -5,6 +5,7 @@ import { MockLanguageModelV3 } from '../test/mock-language-model-v3';
 import { ToolLoopAgent } from './tool-loop-agent';
 import { AsyncIterableStream } from '../util/async-iterable-stream';
 import { DeepPartial } from '../util/deep-partial';
+import { ModelMessage } from '../prompt';
 
 describe('ToolLoopAgent', () => {
   describe('generate', () => {
@@ -18,6 +19,32 @@ describe('ToolLoopAgent', () => {
         system: '123',
         prompt: 'Hello, world!',
       });
+    });
+
+    it('should require options when call options are provided', async () => {
+      const agent = new ToolLoopAgent<{ callOption: string }>({
+        model: new MockLanguageModelV3(),
+      });
+
+      expectTypeOf<Parameters<typeof agent.generate>[0]>().toEqualTypeOf<
+        { options: { callOption: string } } & (
+          | { prompt: string | ModelMessage[]; messages?: never }
+          | { messages: ModelMessage[]; prompt?: never }
+        )
+      >();
+    });
+
+    it('should not require options when call options are not provided', async () => {
+      const agent = new ToolLoopAgent({
+        model: new MockLanguageModelV3(),
+      });
+
+      expectTypeOf<Parameters<typeof agent.generate>[0]>().toEqualTypeOf<
+        { options?: never } & (
+          | { prompt: string | ModelMessage[]; messages?: never }
+          | { messages: ModelMessage[]; prompt?: never }
+        )
+      >();
     });
 
     it('should infer output type', async () => {
@@ -51,6 +78,32 @@ describe('ToolLoopAgent', () => {
       });
     });
 
+    it('should require options when call options are provided', async () => {
+      const agent = new ToolLoopAgent<{ callOption: string }>({
+        model: new MockLanguageModelV3(),
+      });
+
+      expectTypeOf<Parameters<typeof agent.stream>[0]>().toEqualTypeOf<
+        { options: { callOption: string } } & (
+          | { prompt: string | ModelMessage[]; messages?: never }
+          | { messages: ModelMessage[]; prompt?: never }
+        )
+      >();
+    });
+
+    it('should not require options when call options are not provided', async () => {
+      const agent = new ToolLoopAgent({
+        model: new MockLanguageModelV3(),
+      });
+
+      expectTypeOf<Parameters<typeof agent.stream>[0]>().toEqualTypeOf<
+        { options?: never } & (
+          | { prompt: string | ModelMessage[]; messages?: never }
+          | { messages: ModelMessage[]; prompt?: never }
+        )
+      >();
+    });
+
     it('should infer output type', async () => {
       const agent = new ToolLoopAgent({
         model: new MockLanguageModelV3(),
@@ -59,7 +112,7 @@ describe('ToolLoopAgent', () => {
         }),
       });
 
-      const streamResult = agent.stream({
+      const streamResult = await agent.stream({
         prompt: 'Hello, world!',
       });