+++ packages/mistral/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/mistral
 
+## 3.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [fca786b]
+  - @ai-sdk/provider-utils@4.0.0-beta.20
+
 ## 3.0.0-beta.20
 
 ### Patch Changes