+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- Updated dependencies [9b83947]
+  - ai@6.0.0-beta.68
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.67
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [fca786b]
+  - ai@6.0.0-beta.66
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [dce4e7b]
   - ai@6.0.0-beta.65