+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/xai
 
+## 3.0.0-beta.23
+
+### Patch Changes
+
+- b9e5b77: feat(xai): Support `parallel_function_calling` provider option for XAI provider
+
+## 3.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies [fca786b]
+  - @ai-sdk/provider-utils@4.0.0-beta.20
+  - @ai-sdk/openai-compatible@2.0.0-beta.21
+
 ## 3.0.0-beta.21
 
 ### Patch Changes